+++ packages/agents/src/utils/index.ts
@@ -1,6 +1,7 @@
 // Utility exports
 export * from './message-converter';
 export * from './logger';
+export * from './simple-logger';
 export * from './validation';
 export * from './errors';